feat: 이미지 업로드 버튼 분리
