+++ ReactNativeStudy/Zero/Week2/camera/src/constants/size.ts
@@ -1,0 +1,1 @@
+export const IMAGE_CROP_SIZE = 367;
